move to_sample_rdd to utils
